refactor modkit-db module by manually specifying the wanted database
