+++ Cargo.toml
@@ -5,6 +5,7 @@ license = "Apache-2.0"
 authors = ["Hypernetix"]
 repository = "https://github.com/hypernetix/hyperspot"
 description = "HyperSpot Server"
+rust-version = "1.92.0"
 
 [profile.dev]
 incremental = true
@@ -275,15 +276,6 @@ parking_lot = "0.12"
 
 # CLI support
 clap = { version = "4.5", features = ["derive"] }
-
-# Database (only for db module)
-sqlx = { version = "0.8", features = [
-    "runtime-tokio-rustls",
-    "postgres",
-    "sqlite",
-    "any",
-    "migrate",
-] }
 
 # Database utilities
 sea-orm = { version = "1.1", default-features = false, features = [
