+++ apps/gts-docs-validator/src/scanner.rs
@@ -39,10 +39,10 @@ fn matches_exclude(path: &Path, exclude_patterns: &[Pattern]) -> bool {
             return true;
         }
         // Also try matching just the file/dir name
-        if let Some(name) = path.file_name() {
-            if pattern.matches(&name.to_string_lossy()) {
-                return true;
-            }
+        if let Some(name) = path.file_name()
+            && pattern.matches(&name.to_string_lossy())
+        {
+            return true;
         }
     }
     false
@@ -51,10 +51,10 @@ fn matches_exclude(path: &Path, exclude_patterns: &[Pattern]) -> bool {
 /// Check if path contains any skip directories
 fn in_skip_dir(path: &Path) -> bool {
     for component in path.components() {
-        if let std::path::Component::Normal(name) = component {
-            if SKIP_DIRS.iter().any(|skip| name.to_string_lossy() == *skip) {
-                return true;
-            }
+        if let std::path::Component::Normal(name) = component
+            && SKIP_DIRS.iter().any(|skip| name.to_string_lossy() == *skip)
+        {
+            return true;
         }
     }
     false
@@ -148,10 +148,10 @@ pub fn find_files(paths: &[PathBuf], exclude: &[String], verbose: bool) -> Vec<P
 /// Check if a matched string is a false positive
 fn is_false_positive(gts_id: &str) -> bool {
     for pattern in FALSE_POSITIVE_PATTERNS {
-        if let Ok(re) = Regex::new(pattern) {
-            if re.is_match(gts_id) {
-                return true;
-            }
+        if let Ok(re) = Regex::new(pattern)
+            && re.is_match(gts_id)
+        {
+            return true;
         }
     }
     false
