+++ libs/modkit-db-macros/Cargo.toml
@@ -5,6 +5,7 @@ edition.workspace = true
 license.workspace = true
 authors.workspace = true
 repository.workspace = true
+rust-version.workspace = true
 description = "Proc-macro derives for modkit-db secure ORM layer"
 
 [lib]
