+++ libs/modkit-db/src/pool_opts.rs
@@ -12,7 +12,9 @@ pub trait ApplyPoolOpts<T> {
 }
 
 #[cfg(feature = "pg")]
-impl ApplyPoolOpts<sqlx::postgres::PgPoolOptions> for sqlx::postgres::PgPoolOptions {
+impl ApplyPoolOpts<sea_orm::sqlx::postgres::PgPoolOptions>
+    for sea_orm::sqlx::postgres::PgPoolOptions
+{
     fn apply(mut self, opts: &ConnectOpts) -> Self {
         if let Some(n) = opts.max_conns {
             self = self.max_connections(n);
@@ -37,7 +39,9 @@ impl ApplyPoolOpts<sqlx::postgres::PgPoolOptions> for sqlx::postgres::PgPoolOpti
 }
 
 #[cfg(feature = "mysql")]
-impl ApplyPoolOpts<sqlx::mysql::MySqlPoolOptions> for sqlx::mysql::MySqlPoolOptions {
+impl ApplyPoolOpts<sea_orm::sqlx::mysql::MySqlPoolOptions>
+    for sea_orm::sqlx::mysql::MySqlPoolOptions
+{
     fn apply(mut self, opts: &ConnectOpts) -> Self {
         if let Some(n) = opts.max_conns {
             self = self.max_connections(n);
@@ -62,7 +66,9 @@ impl ApplyPoolOpts<sqlx::mysql::MySqlPoolOptions> for sqlx::mysql::MySqlPoolOpti
 }
 
 #[cfg(feature = "sqlite")]
-impl ApplyPoolOpts<sqlx::sqlite::SqlitePoolOptions> for sqlx::sqlite::SqlitePoolOptions {
+impl ApplyPoolOpts<sea_orm::sqlx::sqlite::SqlitePoolOptions>
+    for sea_orm::sqlx::sqlite::SqlitePoolOptions
+{
     fn apply(mut self, opts: &ConnectOpts) -> Self {
         if let Some(n) = opts.max_conns {
             self = self.max_connections(n);
