+++ libs/modkit-db/tests/db_generic.rs
@@ -4,7 +4,7 @@
 mod common;
 use anyhow::Result;
 #[cfg(feature = "pg")]
-use sqlx::Row;
+use sea_orm::sqlx::Row;
 
 #[cfg(feature = "sqlite")]
 #[tokio::test]
@@ -47,19 +47,22 @@ async fn run_common_suite(database_url: &str) -> Result<()> {
         modkit_db::DbEngine::Postgres => {
             let pool = db.sqlx_postgres().unwrap();
             // Create a simple test table and verify
-            sqlx::query("CREATE TABLE IF NOT EXISTS test_pg (id SERIAL PRIMARY KEY, name TEXT)")
-                .execute(pool)
-                .await?;
+            sea_orm::sqlx::query(
+                "CREATE TABLE IF NOT EXISTS test_pg (id SERIAL PRIMARY KEY, name TEXT)",
+            )
+            .execute(pool)
+            .await?;
 
-            let result = sqlx::query("INSERT INTO test_pg (name) VALUES ($1) RETURNING id")
-                .bind("test_user")
-                .fetch_one(pool)
-                .await?;
+            let result =
+                sea_orm::sqlx::query("INSERT INTO test_pg (name) VALUES ($1) RETURNING id")
+                    .bind("test_user")
+                    .fetch_one(pool)
+                    .await?;
 
             let id: i32 = result.get("id");
             assert!(id > 0);
 
-            let count: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM test_pg")
+            let count: i64 = sea_orm::sqlx::query_scalar("SELECT COUNT(*) FROM test_pg")
                 .fetch_one(pool)
                 .await?;
             assert_eq!(count, 1);
@@ -68,16 +71,16 @@ async fn run_common_suite(database_url: &str) -> Result<()> {
         modkit_db::DbEngine::MySql => {
             let pool = db.sqlx_mysql().unwrap();
             // Create a simple test table and verify
-            sqlx::query("CREATE TABLE IF NOT EXISTS test_mysql (id INT AUTO_INCREMENT PRIMARY KEY, name TEXT)")
+            sea_orm::sqlx::query("CREATE TABLE IF NOT EXISTS test_mysql (id INT AUTO_INCREMENT PRIMARY KEY, name TEXT)")
                 .execute(pool)
                 .await?;
 
-            sqlx::query("INSERT INTO test_mysql (name) VALUES (?)")
+            sea_orm::sqlx::query("INSERT INTO test_mysql (name) VALUES (?)")
                 .bind("test_user")
                 .execute(pool)
                 .await?;
 
-            let count: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM test_mysql")
+            let count: i64 = sea_orm::sqlx::query_scalar("SELECT COUNT(*) FROM test_mysql")
                 .fetch_one(pool)
                 .await?;
             assert_eq!(count, 1);
@@ -86,18 +89,18 @@ async fn run_common_suite(database_url: &str) -> Result<()> {
         modkit_db::DbEngine::Sqlite => {
             let pool = db.sqlx_sqlite().unwrap();
             // Create a simple test table and verify
-            sqlx::query(
+            sea_orm::sqlx::query(
                 "CREATE TABLE IF NOT EXISTS test_sqlite (id INTEGER PRIMARY KEY, name TEXT)",
             )
             .execute(pool)
             .await?;
 
-            sqlx::query("INSERT INTO test_sqlite (name) VALUES (?)")
+            sea_orm::sqlx::query("INSERT INTO test_sqlite (name) VALUES (?)")
                 .bind("test_user")
                 .execute(pool)
                 .await?;
 
-            let count: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM test_sqlite")
+            let count: i64 = sea_orm::sqlx::query_scalar("SELECT COUNT(*) FROM test_sqlite")
                 .fetch_one(pool)
                 .await?;
             assert_eq!(count, 1);
