+++ libs/modkit-db/tests/mod.rs
@@ -1,0 +1,3 @@
+//! Integration tests for modkit-db
+
+pub mod sqlite;
