+++ libs/modkit-errors-macro/Cargo.toml
@@ -5,6 +5,8 @@ edition.workspace = true
 license.workspace = true
 authors.workspace = true
 repository.workspace = true
+rust-version.workspace = true
+description = "ModKit errors macro"
 
 [lib]
 proc-macro = true
