+++ libs/modkit-node-info/Cargo.toml
@@ -5,6 +5,8 @@ edition.workspace = true
 license.workspace = true
 authors.workspace = true
 repository.workspace = true
+rust-version.workspace = true
+description = "ModKit node info library"
 
 [lib]
 name = "modkit_node_info"
