+++ libs/modkit-odata/Cargo.toml
@@ -5,6 +5,8 @@ edition.workspace = true
 license.workspace = true
 authors.workspace = true
 repository.workspace = true
+rust-version.workspace = true
+description = "ModKit OData library"
 
 [lib]
 name = "modkit_odata"
