+++ libs/modkit-utils/Cargo.toml
@@ -5,7 +5,8 @@ edition.workspace = true
 license.workspace = true
 authors.workspace = true
 repository.workspace = true
-description.workspace = true
+rust-version.workspace = true
+description = "ModKit utils library"
 
 [lib]
 name = "modkit_utils"
