+++ modules/system/grpc_hub/Cargo.toml
@@ -4,13 +4,14 @@ version.workspace = true
 edition.workspace = true
 license.workspace = true
 authors.workspace = true
-
-[lints]
-workspace = true
+rust-version.workspace = true
+description = "gRPC Hub module"
 
 [lib]
 name = "grpc_hub"
-path = "src/lib.rs"
+
+[lints]
+workspace = true
 
 [dependencies]
 modkit = { workspace = true }
