+++ modules/system/module_orchestrator/module_orchestrator-sdk/Cargo.toml
@@ -4,6 +4,8 @@ version = "0.1.0"
 edition.workspace = true
 license.workspace = true
 authors.workspace = true
+repository.workspace = true
+rust-version.workspace = true
 description = "Domain contracts and client interfaces for module orchestrator"
 
 [dependencies]
