+++ modules/system/nodes_registry/nodes_registry-sdk/Cargo.toml
@@ -1,9 +1,11 @@
 [package]
 name = "nodes_registry-sdk"
-version.workspace = true
+version = "0.1.0"
 edition.workspace = true
 license.workspace = true
 authors.workspace = true
+repository.workspace = true
+rust-version.workspace = true
 description = "SDK for nodes_registry module: API trait, node models, and error definitions"
 
 [lints]
