+++ modules/system/types-registry/types-registry-sdk/Cargo.toml
@@ -1,9 +1,11 @@
 [package]
 name = "types-registry-sdk"
-version.workspace = true
+version = "0.1.0"
 edition.workspace = true
 license.workspace = true
 authors.workspace = true
+repository.workspace = true
+rust-version.workspace = true
 description = "SDK for types-registry module: API trait, GTS entity types, and error definitions"
 
 [lints]
